# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "X API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/media/upload-url": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mimeType": {
                    "type": "string",
                    "enum": [
                      "image/jpeg",
                      "image/png",
                      "image/webp",
                      "image/avif",
                      "image/heic"
                    ]
                  }
                },
                "required": [
                  "mimeType"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "mediaId": {
                          "type": "string",
                          "pattern": "^\\d+$"
                        },
                        "uploadUrl": {
                          "type": "string",
                          "format": "uri"
                        }
                      },
                      "required": [
                        "mediaId",
                        "uploadUrl"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/media/{id}/finalize": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "pattern": "^\\d+$"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "ready",
                            "failed"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "status"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "BLOCKED_BY_USER",
                            "NOT_FOUND",
                            "CONFLICT",
                            "UNPROCESSABLE",
                            "RATE_LIMIT_EXCEEDED",
                            "INTERNAL_ERROR",
                            "SERVICE_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1"
    }
  ]
}